backend/rhythmix_music/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==65.0.2
django-cors-headers==3.13.0
django-extensions==3.2.3
django-oauth-toolkit==3.0.1
django-rest-framework-social-oauth2==1.2.0
django-shell-plus==1.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.10
jwcrypto==1.5.6
multidict==6.1.0
oauth2_provider==0.0
oauthlib==3.2.2
pillow==10.4.0
propcache==0.2.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python3-openid==3.2.0
redis==5.1.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.1.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
spotipy==2.24.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.15.2
